changes double to flato

+++ generate.c
@@ -7,13 +7,13 @@ Particle *generate_bodies(unsigned int N, unsigned int seed) {
     srand(seed);
 
     for (int i = 0; i < N; i++) {
-        particles[i].x = ((double) rand()) / RAND_MAX;
-        particles[i].y = ((double) rand()) / RAND_MAX;
-        particles[i].z = ((double) rand()) / RAND_MAX;
-        particles[i].vx = ((double) rand()) / RAND_MAX;
-        particles[i].vy = ((double) rand()) / RAND_MAX;
-        particles[i].vz = ((double) rand()) / RAND_MAX;
-        particles[i].mass = (1 * ((double) rand()) / RAND_MAX) + 0.001;
+        particles[i].x = ((float) rand()) / RAND_MAX;
+        particles[i].y = ((float) rand()) / RAND_MAX;
+        particles[i].z = ((float) rand()) / RAND_MAX;
+        particles[i].vx = ((float) rand()) / RAND_MAX;
+        particles[i].vy = ((float) rand()) / RAND_MAX;
+        particles[i].vz = ((float) rand()) / RAND_MAX;
+        particles[i].mass = (1 * ((float) rand()) / RAND_MAX) + 0.001;
     }
 
     return particles;
